Fix a bunch of C4267 warnings.
We're still a mess, but slowly chew through it all.

Bug: 22

+++ crypto/bio/bio_mem.c
@@ -142,7 +142,7 @@ static int mem_read(BIO *bio, char *out, int outl) {
   BIO_clear_retry_flags(bio);
   ret = outl;
   if (b->length < INT_MAX && ret > (int)b->length) {
-    ret = b->length;
+    ret = (int)b->length;
   }
 
   if (ret > 0) {
