+++ crypto/fipsmodule/bn/bytes.c
@@ -134,7 +134,8 @@ BIGNUM *BN_le2bn(const uint8_t *in, size_t len, BIGNUM *ret) {
     BN_free(bn);
     return NULL;
   }
-  ret->top = num_words;
+  /* bn_wexpand ensures |num_words| fits. */
+  ret->top = (int)num_words;
 
   /* Make sure the top bytes will be zeroed. */
   ret->d[num_words - 1] = 0;
